Fix bug for Start Acquisition.
The function prototype for CodaRTnetContinuousTracker::StartAcquisition
was incorrectly declared with 'double duration' instead of 'float
duration'. This meant that the method did not overload the method from
the base class, causing one or the other to be called depending on the
type of the duration paramater.

+++ Source Code/Trackers/CodaRTnetContinuousTracker.cpp
@@ -46,7 +46,7 @@ void CodaRTnetContinuousTracker::StartContinuousAcquisition( void ) {
 	cl.startAcqContinuous();
 }
 
-void CodaRTnetContinuousTracker::StartAcquisition( double duration ) {
+void CodaRTnetContinuousTracker::StartAcquisition( float duration ) {
 	for ( int unit = 0; unit < nUnits; unit++ ) nFramesPerUnit[unit] = 0;
 	nFrames = 0;
 	overrun = false;
